Initial commit
I should have committed earlier than this. x_x

Mostly works! It displays the menu, reacts to inputs, and displays the
random value.

I still want to use long-presses to allow you to scroll through history
and do a quick reset.

I still want to add a vibration when the dice are rolled.

I need to update the maifest to have a custom image.